Pkg cleanup: moving out stubs, and independent test packages from pkg/ to test/
