+++ test/stubs/client_stub.go
