+++ test/stubs/netclient_stub.go
